quotes-pipeline/quote_fetcher/requirements.txt
google-cloud-pubsub
requests>=2.26.0,<2.27.0
beautifulsoup4>=4.9.3,<4.10.0
pydantic>=1.8.2,<1.9.0
google-cloud-language>=2.2.2,<2.3.0
